refactor: refactor Upcoming Courses
